Simplify isEqual
