+++ packages/earljs/src/isEqual/index.ts
@@ -1,2 +1,4 @@
 export * from './EqualityOptions'
+export { CanonicalType, getCanonicalType } from './getCanonicalType'
 export { isEqual } from './isEqual'
+export { getKeys, isEqualUnknown } from './isEqualUnknown'
